move repositories to packages
