--- src/main/java/com/social_media/social_media/repository/IPostRepository.java
@@ -1,4 +1,0 @@
-package com.social_media.social_media.repository;
-
-public interface IPostRepository {
-}
