--- src/main/java/com/social_media/social_media/repository/PostRepositoryImpl.java
@@ -1,7 +1,0 @@
-package com.social_media.social_media.repository;
-
-import org.springframework.stereotype.Repository;
-
-@Repository
-public class PostRepositoryImpl implements IPostRepository{
-}
