+++ src/main/java/com/social_media/social_media/repository/follow/FollowRepositoryImpl.java
@@ -1,0 +1,5 @@
+package com.social_media.social_media.repository.follow;
+
+public class FollowRepositoryImpl implements IFollowRepository {
+
+}
