+++ src/main/java/com/social_media/social_media/repository/follow/IFollowRepository.java
@@ -1,0 +1,5 @@
+package com.social_media.social_media.repository.follow;
+
+public interface IFollowRepository {
+
+}
