+++ src/main/java/com/social_media/social_media/repository/post/IPostRepository.java
@@ -1,0 +1,4 @@
+package com.social_media.social_media.repository.post;
+
+public interface IPostRepository {
+}
