+++ src/main/java/com/social_media/social_media/repository/post/PostRepositoryImpl.java
@@ -1,0 +1,7 @@
+package com.social_media.social_media.repository.post;
+
+import org.springframework.stereotype.Repository;
+
+@Repository
+public class PostRepositoryImpl implements IPostRepository {
+}
