+++ src/main/java/com/social_media/social_media/repository/user/IUserRepository.java
@@ -1,0 +1,4 @@
+package com.social_media.social_media.repository.user;
+
+public interface IUserRepository {
+}
