+++ src/main/java/com/social_media/social_media/repository/user/UserRepositoryImpl.java
@@ -1,0 +1,7 @@
+package com.social_media.social_media.repository.user;
+
+import org.springframework.stereotype.Repository;
+
+@Repository
+public class UserRepositoryImpl implements IUserRepository {
+}
