+++ src/main/java/com/social_media/social_media/service/UserServiceImpl.java
@@ -1,11 +1,12 @@
 package com.social_media.social_media.service;
 
-import com.social_media.social_media.repository.IUserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import com.social_media.social_media.repository.user.IUserRepository;
+
 @RequiredArgsConstructor
 @Service
-public class UserServiceImpl implements IUserService{
+public class UserServiceImpl implements IUserService {
     private final IUserRepository userRepository;
 }
